+++ apps/fluent-tester/src/FluentTester/TestComponents/MenuButton/NestedMenuButtonTest.tsx
@@ -8,13 +8,13 @@ export const NestedMenuButton: React.FunctionComponent = () => {
   const [lastMenuItemClicked, setLastMenuItemClicked] = React.useState(null);
 
   const [focusOnMount, setShouldFocusOnMount] = React.useState(true);
-  const toggleFocusOnMount = React.useCallback((value) => setShouldFocusOnMount(value), [setShouldFocusOnMount]);
+  const toggleFocusOnMount = React.useCallback(value => setShouldFocusOnMount(value), [setShouldFocusOnMount]);
 
   const [focusOnContainer, setShouldFocusOnContainer] = React.useState(false);
-  const toggleFocusOnContainer = React.useCallback((value) => setShouldFocusOnContainer(value), [setShouldFocusOnContainer]);
+  const toggleFocusOnContainer = React.useCallback(value => setShouldFocusOnContainer(value), [setShouldFocusOnContainer]);
 
   const onItemClick = React.useCallback(
-    (key) => {
+    key => {
       setLastMenuItemClicked(key);
     },
     [setLastMenuItemClicked],
@@ -38,7 +38,7 @@ export const NestedMenuButton: React.FunctionComponent = () => {
   }, [setShowSubmenu]);
 
   const onSubmenuItemClick = React.useCallback(
-    (key) => {
+    key => {
       setSubmenuLastItemClicked(key);
     },
     [setSubmenuLastItemClicked],