+++ apps/macos/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @fluentui-react-native/tester-macos
 
-This log was last generated on Fri, 10 Sep 2021 21:30:12 GMT and should not be manually modified.
+This log was last generated on Tue, 14 Sep 2021 19:31:04 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.0.1
+
+Tue, 14 Sep 2021 19:31:04 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/tester to v0.43.2 (ruaraki@microsoft.com)
+
+## 0.0.1
+
+Mon, 13 Sep 2021 23:22:43 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/tester to v0.43.1 (ruaraki@microsoft.com)
 
 ## 0.0.1