+++ packages/components/text/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @fluentui-react-native/text
 
-This log was last generated on Fri, 10 Sep 2021 18:16:59 GMT and should not be manually modified.
+This log was last generated on Mon, 13 Sep 2021 23:22:43 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.10.17
+
+Mon, 13 Sep 2021 23:22:43 GMT
+
+### Patches
+
+- Bump @uifabricshared/foundation-compose to v1.10.15 (ruaraki@microsoft.com)
 
 ## 0.10.16